requirements.txt
flask==3.0.*
flask-socketio==5.*
flask-cors==4.*
mediapipe==0.10.14
opencv-python==4.9.*
pygame==2.5.*
elevenlabs==1.*
google-generativeai==0.5.*
pymongo==4.6.*
requests==2.31.*
numpy>=1.26,<2
python-dotenv==1.*
